Moved the handlers so cleaner code
